cowolf (copy-on-write optimization for large files) feature added.
If a file is large and lower branch is read-only, instead of
copying the whole file, this feature saves newly written data in
a sparse file of top branch. It also maintains <offset, len> map
table for written data. If map table entry is available for a read
request, data is served from top branch. If there is no map entry
for a read request, data is read from the lower branch.

Support only two-level branch.